pyproject.toml
[tool.poetry]
name = "twitter-fizzbuzz"
version = "0.1.0"
description = "A Twitter FizzBuzz app"
authors = ["Marcus Gabriel <marcusgabriel.ds@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1"
flask-sqlalchemy = "^2.4"
requests_oauthlib = "^1.3"
python-decouple = "^3.3"

[tool.poetry.dev-dependencies]
pytest = "^5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
